[refactor] Moved server stuff to `net/`.
