--- publish.sh
@@ -1,2 +1,0 @@
-rsync -az data grid.ls.js crossword.ls.js crossword.html bower_components shachari@login.csail.mit.edu:public_html/crossword/
-
